+++ resources/js/components/enso/vuedatatable/VueTable.vue
@@ -10,9 +10,9 @@
             @update-length="length=$event"
             @export-data="exportData"
             @action="action"
-            @reload="getData()"
+            @reload="fetch()"
             @reset="resetPreferences"
-            @request-full-info="forceInfo = true; getData()"
+            @request-full-info="forceInfo = true; fetch()"
             v-on="$listeners"
             v-model="search"/>
         <div class="table-responsive"
@@ -22,7 +22,7 @@
                 id="id">
                 <table-header :template="template"
                     :i18n="i18n"
-                    @sort-update="getData"
+                    @sort-update="fetch"
                     @select-page="selectPage"
                     ref="header"
                     v-if="hasContent"/>
@@ -73,7 +73,7 @@
             :start="start"
             :length="length"
             :selected="selected"
-            @jump-to="start = $event; getData()"
+            @jump-to="start = $event; fetch()"
             v-if="hasContent"/>
         <div class="has-text-centered no-records-found"
             v-if="isEmpty">
@@ -243,7 +243,7 @@ export default {
                 this.template = data.template;
                 this.start = 0;
                 [this.length] = this.template.lengthMenu;
-                this.getData = debounce(this.getData, this.template.debounce);
+                this.fetch = debounce(this.fetch, this.template.debounce);
                 this.setPreferences();
 
                 this.$nextTick(() => {
@@ -251,7 +251,7 @@ export default {
                     this.$emit('initialised');
                 });
 
-                this.getData();
+                this.fetch();
             }).catch((error) => {
                 const { status, data } = error.response;
 
@@ -307,7 +307,7 @@ export default {
             this.search = '';
             this.init();
         },
-        getData() {
+        fetch() {
             this.loading = true;
             this.expanded = [];
 
@@ -320,7 +320,7 @@ export default {
 
                 if (data.data.length === 0 && this.start > 0) {
                     this.start = 0;
-                    this.getData();
+                    this.fetch();
                     return;
                 }
 
@@ -443,7 +443,7 @@ export default {
         ajax(method, path, postEvent) {
             axios[method.toLowerCase()](path).then(({ data }) => {
                 this.$toastr.success(data.message);
-                this.getData();
+                this.fetch();
 
                 if (postEvent) {
                     this.$emit(postEvent);
@@ -468,7 +468,7 @@ export default {
             }
 
             this.start = 0;
-            this.getData();
+            this.fetch();
         },
         selectPage(state) {
             this.$refs.body.selectPage(state);