+++ resources/js/components/enso/vueforms/EnsoForm.vue
@@ -4,8 +4,8 @@
         :locale="locale"
         :params="params"
         :path="path"
-        @loaded="ready = true; $emit('loaded')"
         v-on="$listeners"
+        @loaded="ready = true; $emit('loaded')"
         ref="form">
         <template v-for="field in customFields"
             :slot="field.name"
@@ -26,25 +26,25 @@
 import VueForm from '../../../components/enso/vueforms/VueForm.vue';
 
 export default {
-    name: 'VueFormSs',
+    name: 'EnsoForm',
 
     components: { VueForm },
 
     props: {
-        path: {
+        locale: {
             type: String,
             default() {
-                return route(this.$route.name, this.$route.params, false);
+                return this.$store.state.preferences.global.lang;
             },
         },
         params: {
             type: Object,
             default: null,
         },
-        locale: {
+        path: {
             type: String,
             default() {
-                return this.$store.state.preferences.global.lang;
+                return route(this.$route.name, this.$route.params, false);
             },
         },
     },
@@ -54,26 +54,36 @@ export default {
     }),
 
     computed: {
+        data() {
+            return this.ready
+                ? this.$refs.form.data
+                : null;
+        },
         customFields() {
             return this.ready
-                ? this.$refs.form.data.sections
-                    .reduce((fields, section) => fields
-                        .concat(section.fields.filter(field => field.meta.custom)), [])
+                ? this.$refs.form.customFields
                 : [];
         },
     },
 
     methods: {
         field(field) {
-            return this.ready && this.$refs.form.field(field);
+            return this.ready
+                ? this.$refs.form.field(field)
+                : null;
         },
         param(param) {
-            return this.ready && this.$refs.form.param(param);
+            return this.ready
+                ? this.$refs.form.param(param)
+                : null;
         },
         routeParam(param) {
-            return this.ready && this.$refs.form.routeParam(param);
+            return this.ready
+                ? this.$refs.form.routeParam(param)
+                : null;
         },
     },
 
 };
+
 </script>