+++ resources/js/components/enso/vueforms/VueForm.vue
@@ -1,6 +1,6 @@
 <template>
 
-    <div>
+    <div v-if="data">
         <form-header :data="data"
             :i18n="i18n"/>
         <form class="is-marginless"
@@ -42,6 +42,7 @@
             <form-actions :data="data"
                 :errors="errors"
                 :i18n="i18n"
+                :params="params"
                 v-on="$listeners"
                 @focus-error="focusError"
                 ref="actions">
@@ -72,48 +73,72 @@ export default {
     },
 
     props: {
-        data: {
-            type: Object,
-            required: true,
-        },
         i18n: {
             type: Function,
-            default(key) {
-                return Object.keys(this.$options.methods).includes('__')
-                    ? this.__(key)
-                    : key;
-            },
-        },
-        params: {
-            type: Object,
-            default: null,
+            default: key => key,
         },
         locale: {
             type: String,
             default: 'en',
         },
+        params: {
+            type: Object,
+            default: null,
+        },
+        path: {
+            type: String,
+            required: true,
+        },
     },
 
-    data() {
-        return {
-            errors: new Errors(),
-        };
+    data: () => ({
+        data: null,
+        errors: new Errors(),
+    }),
+
+    computed: {
+        customFields() {
+            return this.data
+                ? this.data.sections
+                    .reduce((fields, section) => fields
+                        .concat(section.fields.filter(field => field.meta.custom)), [])
+                : [];
+        },
+    },
+
+    created() {
+        this.fetch();
     },
 
     methods: {
+        fetch() {
+            axios.get(this.path, { params: this.params })
+                .then(({ data }) => {
+                    this.data = data.form;
+                    this.$emit('loaded');
+                }).catch(error => this.handleError(error));
+        },
         columnSize(columns) {
             return `is-${parseInt(12 / columns, 10)}`;
         },
         flatten() {
             return this.data.sections
-                .reduce((fields, section) => fields.concat(section.fields), []);
+                .reduce((fields, section) =>
+                    fields.concat(section.fields), []);
         },
         field(field) {
             return this.flatten()
                 .find(item => item.name === field);
         },
+        param(param) {
+            return this.data.params[param];
+        },
+        routeParam(param) {
+            return this.data.routeParams[param];
+        },
         hasFields(section) {
-            return section.fields.find(field => !field.meta.hidden) !== undefined;
+            return !!section.fields
+                .find(field => !field.meta.hidden);
         },
         focusError() {
             this.$el.querySelector('.help.is-danger')