+++ resources/js/components/enso/vueforms/parts/FormActions.vue
@@ -117,6 +117,10 @@ export default {
             type: Function,
             required: true,
         },
+        params: {
+            type: Object,
+            default: null,
+        },
     },
 
     data: () => ({