+++ resources/js/pages/administration/people/Edit.vue
@@ -2,15 +2,15 @@
 
     <div class="columns is-centered">
         <div class="column is-three-quarters-desktop is-full-touch">
-            <vue-form-ss class="box has-background-light raises-on-hover animated fadeIn"
-                @loaded="initialised = true"
+            <enso-form class="box has-background-light raises-on-hover animated fadeIn"
+                @loaded="ready = true"
                 ref="form">
                 <span slot="actions">
                     <a class="button is-warning"
-                        v-if="initialised && $refs.form.data.params.userId"
+                        v-if="ready && $refs.form.param('userId')"
                         @click="$router.push({
                             name: 'administration.users.edit',
-                            params: { user: $refs.form.data.params.userId }
+                            params: { user: $refs.form.param('userId') }
                         })">
                         <span class="is-hidden-mobile">
                             {{ __('Edit User') }}
@@ -21,11 +21,11 @@
                         <span class="is-hidden-mobile"/>
                     </a>
                     <a class="button is-primary"
-                        v-else-if="initialised">
+                        v-else-if="ready">
                         <span class="is-hidden-mobile"
                             @click="$router.push({
                                 name: 'administration.users.create',
-                                params: { person: $refs.form.data.routeParams.person }
+                                params: { person: $refs.form.routeParam('person') }
                             })">
                             {{ __('Create User') }}
                         </span>
@@ -35,10 +35,10 @@
                         <span class="is-hidden-mobile"/>
                     </a>
                 </span>
-            </vue-form-ss>
-            <div v-if="initialised">
+            </enso-form>
+            <div v-if="ready">
                 <accessories type="LaravelEnso\People\app\Models\Person"
-                    :id="$refs.form.data.routeParams.person"
+                    :id="$refs.form.routeParam('person')"
                     addresses/>
             </div>
         </div>
@@ -50,16 +50,16 @@
 
 import { library } from '@fortawesome/fontawesome-svg-core';
 import { faUser } from '@fortawesome/free-solid-svg-icons';
-import VueFormSs from '../../../components/enso/vueforms/VueFormSs.vue';
+import EnsoForm from '../../../components/enso/vueforms/EnsoForm.vue';
 import Accessories from '../../../components/enso/bulma/Accessories.vue';
 
 library.add(faUser);
 
 export default {
-    components: { VueFormSs, Accessories },
+    components: { EnsoForm, Accessories },
 
     data: () => ({
-        initialised: false,
+        ready: false,
     }),
 };