+++ resources/js/pages/administration/users/Create.vue
@@ -2,8 +2,7 @@
 
     <div class="columns is-centered">
         <div class="column is-three-quarters-desktop is-full-touch">
-            <vue-form-ss class="box has-background-light raises-on-hover animated fadeIn"
-                :route-params="[$route.name, $route.params.person, false]">
+            <enso-form class="box has-background-light raises-on-hover animated fadeIn">
                 <template slot="group_id"
                     slot-scope="{ field, errors, i18n }">
                     <select-field :errors="errors"
@@ -18,7 +17,7 @@
                         :i18n="i18n"
                         :pivot-params="pivotParams"/>
                 </template>
-            </vue-form-ss>
+            </enso-form>
         </div>
     </div>
 
@@ -26,12 +25,12 @@
 
 <script>
 
-import VueFormSs from '../../../components/enso/vueforms/VueFormSs.vue';
+import EnsoForm from '../../../components/enso/vueforms/EnsoForm.vue';
 import InputField from '../../../components/enso/vueforms/fields/InputField.vue';
 import SelectField from '../../../components/enso/vueforms/fields/SelectField.vue';
 
 export default {
-    components: { VueFormSs, InputField, SelectField },
+    components: { EnsoForm, InputField, SelectField },
 
     data: () => ({
         pivotParams: { userGroups: { id: null } },