+++ resources/js/pages/system/menus/Edit.vue
@@ -2,7 +2,7 @@
 
     <div class="columns is-centered">
         <div class="column is-three-quarters-desktop is-full-touch">
-            <vue-form-ss class="box has-background-light raises-on-hover animated fadeIn"/>
+            <enso-form class="box has-background-light raises-on-hover animated fadeIn"/>
         </div>
     </div>
 
@@ -10,10 +10,10 @@
 
 <script>
 
-import VueFormSs from '../../../components/enso/vueforms/VueFormSs.vue';
+import EnsoForm from '../../../components/enso/vueforms/EnsoForm.vue';
 
 export default {
-    components: { VueFormSs },
+    components: { EnsoForm },
 };
 
 </script>